backend/requirements.txt
altgraph==0.10.2
backports.shutil-get-terminal-size==1.0.0
bdist-mpkg==0.5.0
bonjour-py==0.3
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
decorator==4.2.1
Django==1.11.10
django-cors-headers==1.2.0
django-environ==0.4.4
django-filter==1.1.0
django-graphql-jwt==0.1.5
django-node==4.0.0
django-social-auth==0.7.28
django-webpack==3.1.1
django-webpack-loader==0.6.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
enum34==1.1.6
functools32==3.2.3.post2
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
httplib2==0.10.3
idna==2.6
ipython-genutils==0.2.0
iso8601==0.1.12
macholib==1.5.1
matplotlib==2.0.0
modulegraph==0.10.4
networkx==1.11
numpy==1.11.2
oauth2==1.9.0.post1
oauthlib==2.0.6
optional-django==0.3.0
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
promise==2.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py2app==0.7.3
Pygments==2.2.0
PyJWT==1.6.0
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
pypng==0.0.18
python-dateutil==1.5
python-openid==2.2.5
pytz==2013.7
requests==2.18.4
requests-oauthlib==0.8.0
Rx==1.6.1
scandir==1.7
scipy==0.13.0b1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
subprocess32==3.5.0rc1
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
webpack==6.0.0
xattr==0.6.4
zope.interface==4.1.1
